refactor packages and modid
Renamed everything cloth related to cotton
changed packages (we don't own cottonmc.net) to io.github.cottonmc
