--- src/main/resources/assets/cloth/lang/en_us.json
@@ -1,3 +1,0 @@
-{
-  "itemGroup.cloth:cloth_tab": "Cloth"
-}
