+++ src/main/resources/assets/cotton/lang/en_us.json
@@ -1,0 +1,3 @@
+{
+  "itemGroup.cotton:cotton_tab": "Cotton"
+}
